% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{getErrorsOption}
\alias{getErrorsOption}
\alias{getWarningsOption}
\alias{getWarningsWhitelist}
\title{Helpers for deducing options of mcMap}
\usage{
getErrorsOption(errors = getOption("mctoolsErrors", 0))

getWarningsOption(warn = getOption("warn", 0))

getWarningsWhitelist(whitelist = getOption("mctoolsWarningsWhitelist",
  character()))
}
\arguments{
\item{errors, warn}{(numeric) see \link{options}. '<=1' -> 'suppress';
'>1' -> 'stop'}

\item{whitelist}{(character) see \link{mcMap}}
}
\description{
Helpers for deducing options of mcMap. A translation of R's warning option
into this framework. Logging of warnings and errors is not handled with these
options for \code{mcMap}.
}
