% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcMap.R
\name{WithCallingHandlers}
\alias{WithCallingHandlers}
\title{Error and warning handling}
\usage{
WithCallingHandlers(fun, errors = getErrorsOption(),
  warnings = getWarningsOption(),
  warningsWhitelist = getWarningsWhitelist())
}
\arguments{
\item{fun}{(function) a function}

\item{errors}{(character) one in \code{c("suppress", "log")}}

\item{warnings}{(character) one in \code{c("suppress", "log", "asError")}}

\item{warningsWhitelist}{(character) a vector of regular expressions for white-listing
warnings.}
}
\value{
the function \code{fun} wrapped in calling handlers.
}
\description{
Compared to \link{withCallingHandlers} here we return a function which will
will execute the desired call with modified calling handlers. This exposes
the mechanism used in \link[mctools]{mcMap} to be re-used.
}
\examples{
WithCallingHandlers(function(x) stop("Nooooooo"))()

}
