% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{getErrorsOption}
\alias{getErrorsOption}
\alias{getWarningsOption}
\alias{getWarningsWhitelist}
\title{Deducing options for mcMap}
\usage{
getErrorsOption(errors = getOption("mctoolsErrors", 0))

getWarningsOption(warn = getOption("warn", 0))

getWarningsWhitelist(whitelist = getOption("mctoolsWarningsWhitelist",
  character()))
}
\arguments{
\item{errors, warn}{(numeric) see \link{options}. '-1' -> 'suppress';
'0&1' -> 'log'; '2' -> 'asError'}

\item{whitelist}{(character) see \link{mcMap}}
}
\description{
Helpers for deducing options of mcMap. This is so we can set the behaviour in
the beginning of a script globally. Each call to \code{mcMap} can override
these settings. Logging is enabled by default (for \code{options(warn = 0)}).
For \code{options(warn = -1)} we suppress logging and warnings. For
\code{options(warn = 2)} warnings are handled as errors.
}
